centralize move logic
