+++ core-web/libs/portlets/dot-content-drive/portlet/src/lib/components/dot-content-drive-sidebar/dot-content-drive-sidebar.component.html
@@ -13,4 +13,4 @@
     (onNodeExpand)="onNodeExpand($event)"
     (onNodeCollapse)="onNodeCollapse($event)"
     (uploadFiles)="uploadFiles.emit($event)"
-    (moveItems)="moveItemsToFolder($event)" />
+    (moveItems)="moveItems.emit($event)" />
