+++ core-web/libs/portlets/dot-content-drive/portlet/src/lib/dot-content-drive-shell/dot-content-drive-shell.component.html
@@ -34,7 +34,10 @@
     class="tree-selector center-content"
     [class.expanded]="$treeExpanded()"
     data-testid="tree-selector">
-    <dot-content-drive-sidebar data-testid="sidebar" (uploadFiles)="resolveFilesUpload($event)" />
+    <dot-content-drive-sidebar
+        data-testid="sidebar"
+        (uploadFiles)="resolveFilesUpload($event)"
+        (moveItems)="onMoveItems($event)" />
 </div>
 
 <dot-content-drive-dropzone data-testid="dropzone" (uploadFiles)="resolveFilesUpload($event)">
